# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Confluence Server
host: localhost:10010
basePath: /api/
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /sessions:
    x-swagger-router-controller: sessions
    post:
      description: Login and get UID (set in cookie).
      operationId: login
      parameters:
        - name: login_obj
          in: body
          description: Login object.
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        "403":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /meetings:
    x-swagger-router-controller: meetings
    get:
      description: Get all meetings for the current user.
      operationId: getAllMeetings
      parameters:
        - name: Cookie
          in: header
          description: UID in cookie (leave blank, use sessions api to login beforing using this api)
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MeetingsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Create a new meeting for the current user.
      operationId: createMeeting
      parameters:
        - name: meeting_obj
          in: body
          description: Meeting object.
          required: true
          schema:
            $ref: "#/definitions/MeetingRequest"
        - name: Cookie
          in: header
          description: UID in cookie (leave blank, use sessions api to login beforing using this api)
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /meetings/{id}:
      x-swagger-router-controller: meetings
      get:
        description: Get information for a specific meeting.
        operationId: getMeeting
        parameters:
          - name: id
            in: path
            description: Id of the meeting.
            required: true
            type: string
          - name: Cookie
            in: header
            description: UID in cookie (leave blank, use sessions api to login beforing using this api)
            type: string
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/MeetingResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
      delete:
        description: Delete a specific meeting.
        operationId: deleteMeeting
        parameters:
          - name: id
            in: path
            description: Id of the meeting.
            required: true
            type: string
          - name: Cookie
            in: header
            description: UID in cookie (leave blank, use sessions api to login beforing using this api)
            type: string
            required: true
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/SuccessResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /meetings/suggestions:
    x-swagger-router-controller: meetings
    post:
      description: Get recommendations for a specific meeting.
      operationId: getMeetingSuggestions
      parameters:
        - name: request_obj
          in: body
          required: true
          description: Meeting sugestions request.
          schema:
            $ref: "#/definitions/MeetingSuggestionRequest"
        - name: Cookie
          in: header
          description: UID in cookie (leave blank, use sessions api to login beforing using this api)
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/MeetingSuggestionsResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    x-swagger-router-controller: users
    get:
      description: Get all users.
      operationId: getAllUsers
      parameters:
        - name: Cookie
          in: header
          description: UID in cookie (leave blank, use sessions api to login beforing using this api)
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
#  /rooms:
#    x-swagger-router-controller: rooms
#    get:
#      description: Get all rooms.
#      operationId: getAllRooms
#      parameters:
#        - name: Cookie
#          in: header
#          description: UID in cookie (leave blank, use sessions api to login beforing using this api)
#          type: string
#          required: true
#      responses:
#        "200":
#          description: Success
#          schema:
#            $ref: "#/definitions/RoomsResponse"
#        default:
#          description: Error
#          schema:
#            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  LoginRequest:
    required:
      - username
      - password
    properties:
      username:
        type: string
      password:
        type: string
  MeetingsResponse:
    required:
      - meetings
    properties:
      meetings:
        type: array
        items:
          $ref: "#/definitions/MeetingResponse"
  MeetingRequest:
    required:
      - title
      - note
      - start_time
      - end_time
      - room_id
      - required_ids
      - suggested_ids
    properties:
      title:
        type: string
      note:
        type: string
      start_time:
        type: string
      end_time:
        type: string
      room_id:
        type: integer
      required_ids:
        type: array
        items:
          type: integer
      suggested_ids:
        type: array
        items:
          type: integer
  MeetingResponse:
    required:
      - title
      - note
      - start_time
      - end_time
      - room
      - required_users
      - suggested_users
    properties:
      title:
        type: string
      note:
        type: string
      start_time:
        type: string
      end_time:
        type: string
      room:
        $ref: "#/definitions/RoomResponse"
      required_users:
        type: array
        items:
          $ref: "#/definitions/UserResponse"
      suggested_users:
        type: array
        items:
          $ref: "#/definitions/UserResponse"
  RoomResponse:
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string
  RoomsResponse:
    required:
      - rooms
    properties:
      rooms:
        type: array
        items:
          $ref: "#/definitions/RoomResponse"
  UsersResponse:
    required:
      - users
    properties:
      users:
        type: array
        items:
          $ref: "#/definitions/UserResponse"
  UserResponse:
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string
  MeetingSuggestionsResponse:
    required:
      - conflict
      - suggestions
    properties:
      conflict:
        type: boolean
      suggestions:
        type: array
        items:
          $ref: "#/definitions/MeetingSuggestionResponse"
  MeetingSuggestionResponse:
    required:
      - start_time
      - end_time
      - room
    properties:
      start_time:
        type: string
      end_time:
        type: string
      room:
        $ref: "#/definitions/RoomResponse"
  MeetingSuggestionRequest:
    required:
      - range_start
      - range_end
      - duration
      - required_ids
    properties:
      range_start:
        type: string
      range_end:
        type: string
      duration:
        type: integer
      required_ids:
        type: array
        items:
          type: integer
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string